backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning - Deep Learning
tensorflow==2.20.0
keras==3.0.0

# Machine Learning - Traditional ML
scikit-learn==1.5.0
xgboost==2.1.0

# Computer Vision
opencv-python==4.10.0
opencv-contrib-python==4.10.0
pillow==10.4.0

# Image Processing & Augmentation
albumentations==1.3.1

# Data Processing
numpy==1.26.0
pandas==2.2.0
scipy==1.14.0

# HTTP Client for Roboflow API
httpx==0.27.0

# MongoDB
pymongo==4.6.0
motor==3.3.2
dnspython==2.4.2
certifi==2024.2.2

# Cloud Storage
boto3==1.29.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
firebase-admin==6.5.0

# Image Metadata
exifread==3.0.0

# Data Validation
marshmallow==3.20.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-magic-bin==0.4.14
aiofiles==23.2.1
